move opeation to reducer
